Application now supports sending and recieving
Finished the code on the server end for echoing new messages to all clients. Added code to client end that prompts the client to enter a user name and appends it to their message. Also changed the message board to a scrolling text window instead of a label widget. I have tested it with up to four clients running simiultaneously on my laptop with no issues so far. Would like to add a scroll bar to the message board, and possible clean up the gui formatting tommorow